requirements.txt
fastapi
uvicorn[standard]
streamlit
scikit-learn
torch
torchvision
Pillow
requests
python-multipart
numpy
pandas
matplotlib
jupyter
ipykernel